+++ module-api/src/test/java/org/example/ReservationConcurrencyTest.java
@@ -33,12 +33,21 @@ public class ReservationConcurrencyTest {
 
         List<ReservationSeatDto> reservationSeatDtos = new ArrayList<>();
         reservationSeatDtos.add(new ReservationSeatDto("ROW_A", "COL_1"));
+        reservationSeatDtos.add(new ReservationSeatDto("ROW_A", "COL_2"));
+
+        List<ReservationSeatDto> reservationSeatDtos2 = new ArrayList<>();
+        reservationSeatDtos2.add(new ReservationSeatDto("ROW_A", "COL_2"));
+        reservationSeatDtos2.add(new ReservationSeatDto("ROW_A", "COL_3"));
+
 
         for (long i = 0; i < numberOfThreads; i++) {
             ReservationRequestDto reservationRequestDto = new ReservationRequestDto(i, 2L, reservationSeatDtos);
+            ReservationRequestDto reservationRequestDto2 = new ReservationRequestDto(i+100, 2L, reservationSeatDtos2);
+
             executorService.execute(() -> {
                 try {
                     reservationService.reserveMovie(reservationRequestDto);
+                    reservationService.reserveMovie(reservationRequestDto2);
                 } catch (Exception e) {
                     System.out.println(e.getMessage());
                 } finally {
@@ -51,6 +60,6 @@ public class ReservationConcurrencyTest {
         executorService.shutdown();
 
         List<Long> reservedSeats = reservationSeatRepository.findReservedSeatByScreenScheduleId(2L);
-        Assertions.assertThat(reservedSeats.size()).isEqualTo(1);
+        Assertions.assertThat(reservedSeats.size()).isEqualTo(2);
     }
 }
